requirements-test.txt
# Testes
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.5.0
coverage==7.4.1
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-randomly==3.15.0
pytest-env==1.1.3
pytest-sugar==1.0.0
pytest-clarity==1.0.1
pytest-verbose-parametrize==1.7.0
pytest-instafail==0.5.0
pytest-rerunfailures==12.0
pytest-socket==0.6.0
pytest-freezegun==0.4.2
pytest-mongodb==2.2.0
pytest-redis==4.0.0
pytest-rabbitmq==2.0.0
pytest-kafka==0.1.0
pytest-elasticsearch==0.0.1
pytest-postgresql==4.1.1
pytest-mysql==4.1.1
pytest-sqlalchemy==0.1.0
pytest-django==4.7.0
pytest-flask==1.3.0
pytest-fastapi==0.1.0
pytest-aiohttp==1.0.4
pytest-tornado==0.8.1
pytest-celery==0.0.0
pytest-docker==2.0.1
pytest-kubernetes==0.3.0
pytest-prometheus==0.1.0
pytest-grafana==0.1.0
pytest-loki==0.1.0
pytest-jaeger==0.1.0
pytest-opentelemetry==0.1.0
pytest-sentry==0.1.0
pytest-datadog==0.1.0
pytest-newrelic==0.1.0
pytest-dynatrace==0.1.0
pytest-appdynamics==0.1.0

# Dependências de Runtime
networkx==3.2.1
redis==5.0.1
websockets==12.0
ujson==5.9.0

# Cobertura
coverage==7.3.2
coverage-badge==1.1.0

# Relatórios
pytest-html-reporter==0.2.9
pytest-metadata==3.1.0
pytest-reportlog==0.3.0
pytest-reportportal==5.3.1

# Mocks e Fixtures
faker==20.1.0
freezegun==1.4.0
responses==0.24.1
vcrpy==4.2.1
httpx==0.25.2
aiohttp==3.9.1

# Utilitários
colorama==0.4.6
rich==13.7.0
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Segurança
bandit==1.7.5

# Linting e Formatação
pylint==3.0.3
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pydocstyle==6.3.0

# Dependências de Teste
langchain==0.1.0
cryptography==42.0.2

# Dependências de Monitoramento
psutil==5.9.8

# Dependências de Infraestrutura
networkx==3.2.1
redis==5.0.1
websockets==12.0
ujson==5.9.0
msgpack==1.0.7

# Dependências de Cobertura
coverage==7.3.2
coverage-badge==1.1.0

# Dependências de Relatórios
pytest-html-reporter==0.2.9
pytest-metadata==3.1.0
pytest-reportlog==0.3.0
pytest-reportportal==5.3.1

# Dependências de Mock e Fixtures
faker==20.1.0
freezegun==1.4.0
responses==0.24.1
vcrpy==4.2.1
httpx==0.25.2
aiohttp==3.9.1

# Dependências de Utilidades
colorama==0.4.6
rich==13.7.0
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Dependências de Segurança
bandit==1.7.5

# Dependências de Linting
pylint==3.0.3
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pydocstyle==6.3.0

# Dependências de Machine Learning
scikit-learn==1.4.1.post1

# Dependências de IA
langchain==0.1.0
cryptography==42.0.2

# Web e Dados
flask==3.0.2
pandas==2.2.1 